src/app/components/header/header.component.ts
import { Component, OnInit } from '@angular/core';
import { AuthenticationService } from 'src/app/services';
import { User } from 'src/app/models';
import { Subscription } from 'rxjs';

@Component({
  selector: 'app-header',
  templateUrl: './header.component.html',
  styleUrls: ['./header.component.css']
})
export class HeaderComponent implements OnInit {
  currentUser: User;
  currentUserSubscription: Subscription;

  constructor(private authenticationService: AuthenticationService) { }

  ngOnInit() {
    this.currentUserSubscription = this.authenticationService.currentUser.subscribe(user => {
      this.currentUser = user;
    });

    console.log(this.currentUser);
  }
}
